- added "useFastMode" flag to MerinoCore.GetDataFromFile() which skips the slow part of GetData (the AssetData.ImportAsset() method) and assumes the textAsset is up to date... because it would be, right? MerinoYarnEditorWindow should already be taking care of that!... see MerinoNodeMapWindow.MaybeInit() for example usage - moved some data-related functions from MerinoYarnEditorWindow to MerinoData, and made them into static methods... (I was using MerinoData.GetDefaultData() to test some stuff in the nodemap)
